refactor and start flow gen
