+++ animal_soup/extensions/__init__.py
@@ -1,6 +1,8 @@
 from .behavior_extensions import BehaviorDataFrameExtension
+from .flow_gen_extensions import FlowGeneratorDataframeExtension
 
 __all__ = [
-    "BehaviorDataFrameExtension"
+    "BehaviorDataFrameExtension",
+    "FlowGeneratorDataframeExtension"
 ]
