+++ animal_soup/extensions/behavior_extensions.py
@@ -1,14 +1,10 @@
 from typing import Union
-from ..batch_utils import get_parent_raw_data_path, validate_path, load_df
-from .._behavior import BehaviorVizContainer
-from .._ethogram import EthogramVizContainer
-from .._ethogram_cleaner import EthogramCleaner
-from .._ethogram_comparison import EthogramComparison
+from ..df_utils import get_parent_raw_data_path, validate_path, load_df
+from ..viewers import *
 import os
 import pandas as pd
 from pathlib import Path
 import shutil
-import time
 import warnings
