+++ animal_soup/viewers/_behavior.py
@@ -7,7 +7,7 @@ from warnings import warn
 from mesmerize_core.arrays import LazyVideo
 from typing import *
 
-from .batch_utils import get_parent_raw_data_path
+from animal_soup.df_utils import get_parent_raw_data_path
 
 from decord import gpu as gpu_context
