+++ animal_soup/viewers/_ethogram_comparison.py
